window/showStatus has more progress params
Summary:
This diff adds more flexibility to the LSP server for displaying progress:

(1) It uses numerator/denominator instead of a fraction. I like integers.

(2) Instead of just a percentage fraction, it can pick an arbitrary markdown string to display. In other words, it's now up to the LSP server to have the freedom to chose a meaningful thing to display, and the wisdom to do so in an approachable and standard manner.

I encountered the need for (2) with flow. During initialization, it has a first phase of "parsing all files" where it doesn't know the total but I still want to see how many files it has done. Then it has a second phase of "typechecking" where it knows the total. I want the yellow tab to have useful information about both phases. It would be misleading to pin them down into just one single cross-phase percentage number.

Reviewed By: kcaze

Differential Revision: D8321495

fbshipit-source-id: 8cc7428d6af719d46485ce076d3fdd03d7f269f9