throw errors when tunnel ports are already being used, and return existing tunnel if it already exists
Summary: This is the first diff towards reimplementing refcounting on tunnels so if there are multiple references to a given tunnel, we don't close it just because one of them closes. In this diff, we do better checking of ports that are already being used for tunnels and throw errors accordingly. We also return an existing tunnel if it's already opened.

Reviewed By: stepanhruda

Differential Revision: D8346483

fbshipit-source-id: e98206857c32ed57ceed70c38afe6d6167ada15e