Add ui-component-tools-common usage to js-imports-server
Summary:
This expands the scope of `js-imports-server`, having it track component definitions. Currently this includes a subset of a component's required props (inferred from Flow typing). It is a subset because this commit does not include more advanced traversal such as looking at imported types and spread type objects.

This commit extends `js-imports-server` in a way that likely wasn't intended. It currently makes no attempt to change the architecture of `js-imports-server` to be more extensible with a greater separation of concerns. As such, the code is a little messy. It's likely that this functionality will motivate (at minimum) the renaming of `js-imports-server` to `js-language-server` as discussed with hansonw and wbinnssmith.

This commit renames the internal `fb-ui-component-tools-common` package to `nuclide-ui-component-tools-common`.

Reviewed By: hansonw

Differential Revision: D8280233

fbshipit-source-id: 058969b445d195b15a3b4ab81c0ab69129b7e77e