Make ChangedFile functionality depend on props
Summary:
ChangedFile made a lot of assumptions about what callbacks would be provided to it, and all of its context menu options were controlled by MultiRootChangedFilesView.

This diff makes the presence of icons and context menu items depend specifically on the props of the ChangedFile component, hopefully making it easier to use this in multiple places.

Reviewed By: evangrayk

Differential Revision: D8319299

fbshipit-source-id: 7c3293caac4c3f0d1a29f44207b3c1b64d385693