Adds traversal of default props to UI Component AST
Summary: ComponentDefinition instances will want to track whether a required prop has a default value or not. This commit introduces an exposed function to accomplish this. It supports a single level of declaration indirection (e.g., in the case of `static defaultProps = DEFAULT_PROPS`). In the future, support for traversing indirection and spreading will either be enhanced or a transition to a `react-docgen` resolver will be made.

Reviewed By: wbinnssmith

Differential Revision: D8406744

fbshipit-source-id: 7fa91572d277874f31bb01166b7a7e4f0730714e