Ensure RpcConnection.dispose() is called in big-dig
Summary:
Shit, `InfoService.shutdown` wasn't completely updated to deal with the big-dig updates - there's definitely some work to be done here to eliminate the usage of `NuclideServer` (it's also being used for shutdown hooks :()

T29368542 tracks cleaning that up, but in the meantime here's a quick fix to ensure that `RpcConnection.dispose()` is always called. (This is actually a pretty big memory leak so big thanks to wanderley for noticing this).

Reviewed By: wanderley

Differential Revision: D8438773

fbshipit-source-id: 41396ee4c94e7f3e95f1549218e608373e529c75