Fix crash on hyperclick with null title
Summary:
The difference between `name?: string` and `name: ?string` is that if you have `foo.name = null` and try to serialize it, nuclide-rpc will fail.

D8324434 recently introduced a new optional field 'title' into the response from textDocument/definition. Flow returns null from it. That was getting put into a Definition object with `definition.name = null`. nuclide-rpc failed to marshal it.

This fixes it, by only having a property `definition.name` if the value is non-null.

Reviewed By: hansonw

Differential Revision: D8464870

fbshipit-source-id: 40d9dbd557573a6f054b1996383abe8affcfe8b8