Settings gear only lists providers for current editor
Summary:
Prior to this diff, when you clicked the settings gear, it'd list ALL providers. e.g. on a JS file, it'd ask you what is your visibility preference for ocaml.

That's not right! because even if you clicked "show always" it still wouldn't be shown!

After this diff, the gear icon only lists providers relevant to the current editor.

(I also sorted the gear tooltip in priority order).

Reviewed By: kcaze

Differential Revision: D8464576

fbshipit-source-id: dc241e2bb3e36bb28bc3f4a6a38e833313e36c3f