LSP allows a project-relative flow binary
Summary:
This diff changes Nuclide's createMultiLspLanguageService to support Flow's desired algorithm for locating the flow binary:
1. See if `<projectRoot>/node_modules/.bin/flow.cmd` exists (if canUseFlowBin and win32)
2. See if `<projectRoot>/node_modules/.bin/flow` exists (if canUseFlowBin>)
3. See if `<pathToFlow>.cmd` exists (if win32)
4. See if `<pathToFlow>` exists

Actually Flow traditionally skipped step 2 in the case of windows, which was inconsistent with the fact that it always did step 4, so I've regularized it.

The first interesting thing is that the choice of paths to search depends upon something only known over RPC (namely, whether it's win32 or not).

The second interesting thing is that if there are relative paths involved, then we can't know at createMultiLspLanguageService time whether any paths will succeed; we'll only know at languageServiceFactory time (i.e. when Nuclide actually demands a specific LspLanguageService from a particular directory).

APPROACH 1:

I think the right long-term approach is to have an active "customization" object which lives on the rpc side, and which is invoked at all stages to customize the language extension. But that's a big refactoring for a future diff. I wanted a smaller change that would work now.

APPROACH 2:

What we've been doing so far is putting everything and the kitchen sink into createMultiLspLanguageService and LspLanguageService - making them general enough to cope with all the needs of all languages. I've continued that trend.

I've made _commands take an array of possible paths.

ODDITY

There's something a bit weird. The historical context is that LspLanguageService is pretty spammy with angry red notification boxes when it can't spin up an LspLanguageService. So hansonw toned that down by saying "if the LSP binary can't be found, then let's have createLspLanguageService simply return null, so language extensions can instantiate a NullLanguageService instead and avoid the spam.

Now to the weird stuff. Prior to this diff, when we call `createMultiLspLanguageService('./flow')` then it will do a screening check whether the `./flow` binary exists. Relative to which path? I don't know but I guess home. Later on, when it's time to actually spawn the LSP server, it's spawned relative to the project directory.

Thus if you had `./flow` in your home but none in the project directory then it would pass the screen but pop up a red notification. Contrariwise if you had none in your home but you did in your project directory then you'd silently get no language services at all.

I think the correct long term future is to remove the optimization. Let's move all language services over to the new 'StatusConfig' that kcaze and I have been writing, and let us always create a LspLanguageService. This way, if something can't be found then it's a very discrete red little box in the corner of your editor, and users are empowered to hide it permanently so they never get spammed again. (Or, if the user does want it, they can simply place the LSP server binary in place and click the 'restart' button. Either way they get better actionable feedback).

But for now, languages haven't momved over to the new StatusConfig. Therefore the optimization has to remain in place. What I've said is: if the set of possible filepaths for the flow binary inclide NO relative paths, then continue to do the optimization; if they do include relative paths then don't.

Reviewed By: hansonw

Differential Revision: D8474646

fbshipit-source-id: 73cf1e95ff0b516339845499923d46237a12cd6a