Change component indexing prototyping script
Summary: There was no reason to have this script be FDS only and there are other namespaces I want to test on. This commit makes the CLI arg be the entire prefix to the glob instead of just a section within a given repository, allowing one to see the `ComponentDefinition` instances indexed for components in any tree.

Differential Revision: D8449316

fbshipit-source-id: 5572ca1abb6fb281b7c96f0575a68ec195ce07bb