Add pattern exclusion to nuclide-projectionist
Summary:
On large repos the rules in a projection config can generate a lot of noise. This commit adds an `exclude` property to `Projection`s. Any path that matches a `fileGlob` but also matches the `exclude` glob will be ignored, e.g., the given `alternate`s will not be shown. Check the tests out for an example.

(NOTE) This does not conflict with vim-projectionist but it is not supported by it.

Reviewed By: wbinnssmith

Differential Revision: D8555298

fbshipit-source-id: fdbfb1c5e33a87417c3cae5801595176a0f58a17