Add nuclide_scribe_macos GK to control scribe logging
Summary:
This is a pre-emptive fix in case we don't figure out what's going on by Monday. (see S161849, https://fb.facebook.com/groups/dirmain/permalink/1956437731071499/)

The GK is currently at 0% while the issue is happening - it may be useful to keep around as a killswitch in case it acts up in the future.

Reviewed By: jgebhardt

Differential Revision: D8593644

fbshipit-source-id: e9d9d3444601bf80e166b2b41757abc773b6c1d0