Fix erroneous association in ComponentDefinition management
Summary: This is an excellent example of what happens when you're not cautious about separating identity values from error! This commit fixes a bug where restoring a `DefinitionManager`'s map of `ComponentDefinition` instances from the cache could result in a component definition being clobbered over with no required props. This occurred because there would be multiple files including the same component name, for instance `FDSButton.react.js` and `FDSButton.react.example.js`. A `ComponentDefinition` for `FDSButton` would be produced from the latter with no requird props and clobber over the instance previously assigned to the former. This commit ensures that definitions are only retrieved from file names in the exact format of `ComponentName.react.js`. It also ensures that if `getRequiredPropsFromAst` fails to find required props that this error state would result in _no_ `ComponentDefinition`, instead of simply an empty array.

Reviewed By: hansonw

Differential Revision: D8595585

fbshipit-source-id: 45ea657b9723bb0d4b87b7d30818d68dc5689d1e