enable react/no-access-state-in-setstate
Summary:
This enables the rule `react/no-access-state-in-setstate` in the Nuclide codebase. It also applies lint suppressions to every current violation so that we can benefit from the rule without changes to the existing code.

Because [`setState` calls are enqueued](https://reactjs.org/docs/react-component.html#setstate), changes are not guaranteed to be applied immediately, and state may have changed in the meantime. The updater function allows React to get the new state at the right time.

Reviewed By: hansonw

Differential Revision: D8585855

fbshipit-source-id: c50d18e5dfe292706413a01096664d32867fb2cb