Fix changing cwd with uncommitted changes crashing smartlog
Summary: I think we switched to using relativePath recently, but when switching cwd it appears that it tries a re-render where filePath is the new uncommitted changes but rootPath is the *old* cwd. This makes `nuclideURI.relative` throw. I'm not sure why we're getting the old cwd, but it's easy enough to just `try/catch` the relativization.

Reviewed By: matthewwithanm

Differential Revision: D8643651

fbshipit-source-id: 445f0668520433cd97aa2ecedfecaa103b914b5f