Adjust layout of buck target box
Summary: The box is not taking up allavailable room anymore. Also, even if it does it's possible that it's not wide enough depending on screen size, length of the buck target string, and length of the device selection strings. Adding a tooltip on hover and having the input box scroll all the way to the right on blur because when there's not enough room, seeing the end of the string (cutting off the beginning) is typically more useful than seeing the beginning (and cutting off the end). Hovering will show the entire string in a tooltip in these cases.

Reviewed By: Goom11

Differential Revision: D8596882

fbshipit-source-id: b59a54c4d457a7c39857167f9da625a7f4919830