Prefer revealing a text editor
Summary:
In D7884681, we added support for revealing ImageEdtior, but it changed the behavior when the active pane was not a text editor.

This takes us back to the old behavior, where you could issue "reveal in file tree" from the keyboard or dash and it would focus the active text editor even if the active pane was, say, File Tree.

Reviewed By: shunfu

Differential Revision: D8652121

fbshipit-source-id: 659e55ea32de2afba5abe02656b8867724536916