Notification for Processes Exited
Summary: A notification is given when a process exits. This will become more useful for multi-targetted debugging. Perhaps it may be less intrusive (and better) to simply write in the console log that a process exited (instead of having a pop-up notification). However, I'm not sure how to do that, or if that's even possible.

Reviewed By: Goom11

Differential Revision: D8628025

fbshipit-source-id: 34fb4aae2838c7999bec09df21b790f05d64d611