Use babel 7.0.0-beta.47 during transpilation
Summary:
This is a big upgrade! But luckily most of the changes were pretty minor.

1) babel-plugin-idx needed an upgrade
2) the transform-modules-commonjs supports laziness now (Unfortunately the output is a bit less efficient - it always requires a function call)
3) inline-invariant-tr needed a slight touchup
4) modules-dependencies parser had to handle babel/xyz
5) `parserOpts` is no longer necessary

Otherwise I just bumped the version and used the equivalent transform from the "babel" namespace - with the exception of `babel/plugin-proposal-class-properties`. Oddly that one now transforms all private class members for some reason.

I enabled the optionalChaining and nullCoalescingOperator plugins as well! So they should be usable now.

here's a diff of transpiled output for `lib/main.js`:

P59768012

Reviewed By: zertosh

Differential Revision: D8673206

fbshipit-source-id: c3690a901fbf769cad376f30fbdf9f4ef2b8025d