Use mousemove instead of mouseenter to select items
Summary:
It seems that if you mount a DOM element with a `mouseenter` listener and your mouse happens to be over that item, the `mouseenter` event will fire immediately.

This is probably good for some cases, but not for quick open (where you often just want to type something and hit enter once the first result is what you want.)

This switches the event listener to `mousemove` which will only change the selection once the mouse is actually moved. (Since this is a recent report this must be Atom 1.28 related?)

`mousemove` is potentially inefficient since it'll fire really often, so I made `QuickSelectionComponent` pure to compensate.

Reviewed By: wbinnssmith

Differential Revision: D8705663

fbshipit-source-id: e7fb7a2b0b8af24cd0cdfca13e974140e0aa8662