Use Mouse Position for Nuclide Rename UI (Context Menu)
Summary:
Right-clicking on a symbol and selecting `Rename` from the Context Menu will select and rename the word at the most recent mouse position (where the right-click occurred), instead of the cursor position.

Note that the UI was already sending the position of the word that the user had right-clicked on (and thus performing the rename correctly). It just so happened that the UI itself was rendering at the word where the cursor was located at.

Reviewed By: nmote

Differential Revision: D8630282

fbshipit-source-id: 8f0da05539060972057ef27f90fae4d8604fe956