Added Big-Dig Thrift Client Manager Class and Tests
Summary:
Added a Big-Dig `ThriftClientManager` class skeleton and tests. This class will be used by `BigDigClient` to manage the thrift service clients, including launch and disposal of Thrift client and the big-dig tunnels they use, etc.

To smooth the review process, I break down the implementation of the `ThriftClientManager` class into several smaller diffs including tests. This diff is the skeleton of the class. Will add the implementation code for each method and tests in later diffs stacked on this one.

I also added a test suite skeleton file for `ThriftClientManager`, will add tests along with the implementation of launch and close logic.

Please help review the structure of class and test file. Thanks!

Reviewed By: semmypurewal

Differential Revision: D8695194

fbshipit-source-id: 69bf6564d6a796b86beaa41650a6659d8352c4ff