Fix max call stack error when revealing in file tree
Summary:
The recursive implementation grew the call stack linearly with the number of siblings we had to traverse. That meant that revealing a file in a folder with a number of children greater than the max stack size could error. This switches to a loop so that we don't have that problem.

Indexing is 1-based in the file tree for reasons I don't understand…I tried to sidestep that confusion by using "offset" here.

Reviewed By: hansonw

Differential Revision: D8794473

fbshipit-source-id: cae6cddb103e29b7a1786a9c4254fb6e7e01d55d