Added onUnexpectedConnectionEnd method to Set Callbacks for Broken Connection
Summary: Before, both lost connection and user closing client events will trigger `onConnectionEnd` and call the same callback function it sets. But the two cases are not quite the same, we should differentiate the two cases therefore in this diffs, I added a new method called `onUnexpectedConnectionEnd` to indicate that the connection `end` event is caused by connection issue and not because users want to close it, which means we probably need to signal users to recreate a new client in the callback function.

Reviewed By: wanderley

Differential Revision: D8797531

fbshipit-source-id: 30629ac48d10fc00a4a11e5c9e7d7ad6d0ad5f1d