Make createMultiLspLanguageService take just a single command not an array
Summary:
createMultiLspLanguageService initially just took just a single command string, for the lsp binary to execute.

We encountered the need with Flow to search for several paths, including ones relative to the project root. A whole lot of infrastructure was built around this to make it work.

But in the end it was a vain hope. Flow's needs are more complex and flow-specific than can be solved with just a list of paths.

So I'm ripping it out.

Reviewed By: mostafaeweda

Differential Revision: D8832151

fbshipit-source-id: 371f8622ac92343afb25cf696b7c371a311312b8