Refactor LspLanguageService command
Summary:
This is a simple refactoring of LspLangugeStatus.start(), to make "getting the command string" its own separate code block, with its own separate error handling. (rather than rolling it into the error handling of the "spawn" command).

The reason is for a future diff, where we'll want to implement more complex algorithms for figuring out the command.

Reviewed By: mostafaeweda

Differential Revision: D8832164

fbshipit-source-id: a3ac5ab3b574c2da4142d291bd7e217d7b8106e5