Implement flow algorithm for finding command
Summary:
Flow has a very specific algorithm it needs to figure out which command to use to spawn the LSP server. (this will be a common need for any project whose configuration files say which version of the toolset to use; it's also done in ocaml and hack, but by the hackier mechanism of relying upon the CWD at the time the lsp binary is launched).

Where to place this algorithm? Given that we've already embraced language-extension-specific code in the common areas (e.g. in ConfigCache search strategy), we're just following precedent by putting this code into LspLanguageService.

Another option would have been to use a language-specific server-side callback to fetch the command. That comes with its weight of extra machinery.

Another option would have been to use a language-specific client-side callback to fetch the command. This would come with its own heavyweight machinery too - (1) a client-side RPCable object just for this single callback, (2) expose a whole host of file-system methods like 'which' and 'exists' and 'runCommand' over RPC.

I'm sure that ultimately we'll need to expose all those file-system methods over bigdig. However, that's a major architectural shift best let for the future when we have more context on the precise needs. And part (1), the client-side RPCable object for this callback, won't be needed at that time. So it felt wrong to built out a shonky version of that infrastructure now.

Note: I passed the flow parameters as a JSON-encoded string in the 'command' parameter. That's pretty ugly. But it's less ugly that polluting the entire callchain with signatures that reveal all the language-specific stuff we're going to be doing. Our end direction is that LspLanguageService should not have language-specific stuff in it, and I think that sticking language-specific signatures into it is a move in the wrong direction.

Reviewed By: mostafaeweda

Differential Revision: D8832211

fbshipit-source-id: 539d38060a1a61d53a7f474eb76d0b48ae700024