Accept FreeformRefactor responses alongside InlineRename
Summary:
After integrating the DiffPreview UI into Refactorizer, it assumed all the edits would be in `TextEdit` format, and it would convert these edits into `ExternalTextEdits` if they happened to apply to multiple files.

However, the `FreeformCompnent` still sends back `ExternalTextEdits` by default, so we needed to take this into account and only convert the edits when necessary.

Also, Flow couldn't associate the types correctly between the response and the type of edit it would contain so there was no choice but to associate each edit explicitly with its type.

In the future: I think we should only have two edit types. But right now the Refactorizer RPC is still sending back edits as absolute character offsets with type `external-edit`. I plan to change the RPC logic in the future so that the format of its responses are aligned with LSP.

Reviewed By: arxanas

Differential Revision: D8849284

fbshipit-source-id: 78e01380c57660eeb272ec569174c6696c1f52ae