Fix potential infinite loop in VSCodeDebuggerAdapterService.getBuckRootFromUri
Summary:
I accidentally passed a non-absolute path when attempting to use the C++ debugger and triggered an infinite loop on the Nuclide server.

Fortunately I was able to attach a debugger and tracked it down to this function; it needs to bail early for non-absolute paths. (Or should this happen on the caller?)

If the path isn't absolute, getting the parent always prepends `../`, which never terminates.

Reviewed By: mostafaeweda

Differential Revision: D8863100

fbshipit-source-id: 49249a7dc1ab9ed2e7c17524a3ecea1d38abde97