Replace former RenameComponent with the InlineRenameComponent
Summary:
Removing the deprecated Rename component of Refactorizer and replacing it with the InlineRenameComponent.

Note that Refactorizer uses `FreeformComponent` for all the refactors that you see me using in the video. I've kept this component because we do use it to extract `async` functions and the like.

After discussing offline with @[1808457483:hchau], I will also in general be removing the refactoring functionality from `nuclide-clang` since they depend largely on logic and formatting that was specific to the old RenameComponent and based on Scuba, no one seems to have used this feature of the package in the past 30 days.

Lastly, I am in general just renaming anything `inline-rename` into `rename`.

Reviewed By: arxanas

Differential Revision: D8849464

fbshipit-source-id: 0e36b1732afa2c9c345a7ca17ab9a549e8aeb84d