Dont force non-optional nullable keys from being present in the original object when marshalling
Summary:
During unmarshalling we're explicitly ignoring keys that are not present in an object if they're nullable because `JSON.parse` that we use for transport has that logic as well.
However, we don't apply the same reasoning to the counterpart function. This makes certain situations fail in nuclide services:

A service that provides: `getDevices(): Array<Device>, queryDevice(device: Device)` where `type Device: {name: ?string}` and with `getDevices` returning `[{name: undefined}]` and then we do `await service.queryDevice(await getDevices()[0])` will fail because the name property was stripped on the return of `getDevices` but when mashalling it to `queryDevice` it would fail.

This scenario was happening with the `AdbService`.

Reviewed By: hansonw

Differential Revision: D8903465

fbshipit-source-id: 3dd7d34450423f143aedda37048d3d408c00354b