Show unreachable breakpoints in their own section in the UI
Summary: This adds some UX for dealing with breakpoints that exist in remote file paths that are unavailable. They will be shown in their own section in the UX, with some text explaining why the breakpoints aren't reachable, and won't be set in the debugger.  This also ensures that if the user clicks on such a breakpoint, we don't try to open a text editor for that path (we currently try, which results in an error message and a black editor tab).

Reviewed By: mostafaeweda

Differential Revision: D8775547

fbshipit-source-id: f5dc99e0bcde30209642f0727cd8caadfe3d3a8e