Remove redundant code to close tunnels
Summary:
The tunnel is closed after all its clients disconnect. Because of that, we don't need to explicity close but wait them to close only after all clients have a chance to proper disconect.

We might need a way to force the disconnection of all thrift server for cases where user closes Nuclide.  Today, this process is done by RPC Connection which for example closes all observables the connection is closed.  To implement something similar, we will need to add support for a close event in the BigDig transport.

Differential Revision: D8920149

fbshipit-source-id: 9da096ef2db747f7c8370df6da466be741141a68