cli: Only print a stop message for the first stop reason
Summary: Adapters can send multiple stop events, one per thread. This makes sense in a GUI debugger where can just update thread states in a window, but in the cli we should only print a stop message on the first stop event to let the user know the debugger is interactive again.

Reviewed By: ebluestein

Differential Revision: D8923803

fbshipit-source-id: 5313fa3e7aa4a2bc7d71394149152f1c6d516c99