cli: breakpoint resolution message on continue displays prompt
Summary:
we stop console input once execution has restarted, but some adapters send output before that happens. since the continued notification is async, the debugger will treat that output as if it happened while
the command prompt is up, and reprint the prompt after it. really, any output that happens while we're trying to continue should see that
input is stopped.

Reviewed By: ebluestein

Differential Revision: D8925121

fbshipit-source-id: 48252fd8c980a96e4996ffa6e9322d2dafb418e3