Store tunnel by service config id
Summary: The current implementation uses only the service name as identifier for a connection to the server.  This is weak because one might want to have multiple servers with different versions.  Before we fix that, I want to refactor ThriftClientManager to cache stuff in terms of an uniquee service config id.

Differential Revision: D8922540

fbshipit-source-id: f554e6ce4830d0ec8a4a6572c74b2fe49c9e33d4